requirements.txt
fastapi
uvicorn[standard]
python-multipart
fastapi-login
passlib[bcrypt]
pycryptodome
python-jose
jinja2
rich
loguru